requirements.txt
dj-database-url==1.0.0
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
requests==2.28.1
whitenoise==6.2.0
django-cors-headers==3.13.0
django-allauth==0.44.0
celery==3.1.18
redis==4.3.4
 
